Move Keychain + GitHubTokenCache to RunnerBarCore (#1608)
- Add Sources/RunnerBarCore/Services/Keychain.swift (public API)
- Add Sources/RunnerBarCore/GitHub/GitHubTokenCache.swift (self-import removed, public)
- Delete Sources/RunnerBar/Services/Keychain.swift
- Delete Sources/RunnerBar/GitHub/GitHubTokenCache.swift

Resolves #1607
